requirements.txt
boto3
tqdm
pyyaml